Train all three learners in full float64 (weights included)
The pipeline runs under jax_enable_x64 (main_unified.py:63), so the data -- plant
integration, buffers, signatures, and every forward/backward computation -- was
already float64. Two float32 surfaces remained, and this closes both.

1. THE INITIAL PATH (finding F-E1). value_gradient_jax._fill_buffer_initial cast the
   value gradient's initial history to float32 while writing, then the float64 buffer
   upcast the already-rounded values back. Commits 07b8d4a/7b1153e moved the buffers to
   float64 and missed this path. Removed (dtype np.float32 -> np.float64). It also made
   the value gradient and the actor-critic start from initial paths differing by
   ~1.2e-8 -- a confound in the learner comparison -- which is now gone.

2. THE NETWORK WEIGHTS. Flax defaults param_dtype to float32 REGARDLESS of x64, and no
   nn.Dense overrode it, so the critic and actor weights of ALL THREE learners were
   float32 even though their inputs and gradients were float64 (float32 params @ float64
   input promotes to a float64 result, which is why the forward pass looked float64 and
   the narrowing was easy to miss). Setting param_dtype=jnp.float64 on every parametric
   layer of the campaign modules (ActorFlax, CriticFlax and their LayerNorm variants)
   makes the weights float64 too.

WHY, for this study specifically: the phenomenon under measurement is the conditioning
of the gradient extraction d_x V. On the high-dimensional platoon cell that passes
through a ~3320-dimensional near-singular readout, where float32 weights can change the
extracted gradient. Studying conditioning in float32 is measuring the quantity with an
instrument that perturbs it. Full float64 is the deliberate, consistent choice.

The change is SYMMETRIC across the three learners, so it adds no confound between them.
It does change absolute results against the pre-2026-07-17 float32-weight runs (all of
main_unified), which remain reproducible from their own commit, not from HEAD.

Verified by effect: all three learners now report critic and actor params float64
(measured on linear_dde_scalar/raw_history), and all three train end-to-end through
main_unified.py to evaluation. Test suite 118 passed, 1 pre-existing failure
(test/study/test_aggregation.py::test_replot_from_aggregation_json_roundtrip).

+++ src/agents/value_gradient_jax.py
@@ -158,18 +158,28 @@ class ContinuousValueGradient:
             
             # Ensure we fill the buffer completely (window_size + 1 elements)
             target_len = self.sliding_signature.window_size + 1
+            # float64, not float32 (finding F-E1, commits 07b8d4a/7b1153e moved the buffers to
+            # float64 and missed this path). The pipeline runs under jax_enable_x64
+            # (main_unified.py:63), and the buffer's own dtype is float64, so the previous float32
+            # cast narrowed ONLY the initial history and then upcast the already-rounded values
+            # back to float64. Keeping it made the value gradient and the actor-critic start from
+            # initial paths differing by ~1.2e-8 (a confound in the learner comparison), and
+            # rounded the very quantity whose conditioning this study measures (raw_history's Gram
+            # on platoon is ~3320-dimensional and near-singular; float32 rounding in the gradient
+            # extraction can change the answer). Removed. The existing main_unified runs predate
+            # this and are reproducible from their own commit, not from HEAD.
             if len(subsampled) >= target_len:
                 # Use the last target_len elements
                 for x in subsampled[-target_len:]:
-                    self.sliding_signature.buffer.append(np.asarray(x/self.training.scale, dtype=np.float32)) #type: ignore
+                    self.sliding_signature.buffer.append(np.asarray(x/self.training.scale, dtype=np.float64)) #type: ignore
             else:
                 # Pad with first element to reach target_len
-                first_val = np.asarray(subsampled[0]/self.training.scale, dtype=np.float32)
+                first_val = np.asarray(subsampled[0]/self.training.scale, dtype=np.float64)
                 padding_needed = target_len - len(subsampled)
                 for _ in range(padding_needed):
                     self.sliding_signature.buffer.append(first_val) #type: ignore
                 for x in subsampled:
-                    self.sliding_signature.buffer.append(np.asarray(x/self.training.scale, dtype=np.float32)) #type: ignore
+                    self.sliding_signature.buffer.append(np.asarray(x/self.training.scale, dtype=np.float64)) #type: ignore
             
             self.sliding_signature.current_signature = self.sliding_signature.compute_signature()
         self._path_data_dirty = True  # Invalidate cache
